requirements.txt
absl-py==1.0.0
astor==0.8.1
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Flask==1.0
gast==0.5.3
grpcio==1.43.0
h5py==2.7.0
idna==2.7
imageio==2.13.5
importlib-metadata==4.8.3
imutils==0.5.3
itsdangerous==2.0.1
Jinja2==3.0.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==2.1.0
networkx==2.5.1
numpy==1.16.2
opencv-python==3.4.3.18
Pillow==8.4.0
protobuf==3.19.3
psutil==5.4.8
pyparsing==3.0.6
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
PyWavelets==1.1.1
pywin32==303
PyYAML==6.0
requests==2.20.0
scikit-image==0.17.2
scikit-learn==0.19.0
scipy==1.2.0
six==1.16.0
tensorboard==1.12.2
tensorflow==1.12.2
termcolor==1.1.0
tifffile==2020.9.3
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchvision==0.11.2+cu113
typing_extensions==4.0.1
urllib3==1.24.3
waitress==1.1.0
Werkzeug==0.15.3
zipp==3.6.0
